backend/requirements.txt
# FastAPI & Server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.30
aiosqlite>=0.20.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Document Parsing
PyPDF2>=3.0.1
python-docx>=1.1.2

# Embeddings
sentence-transformers>=3.0.1

# Vector Databases
chromadb>=0.5.0
faiss-cpu>=1.9.0

# Gemini
google-generativeai>=0.7.2

# Groq (high free-tier limits: 6000 RPM)
groq>=0.9.0

# LangChain
langchain>=0.2.6
langchain-google-genai>=1.0.7
langchain-community>=0.2.6

# Utils
pydantic>=2.10.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.1
httpx>=0.27.0
numpy>=2.0.0
